test: split dashboard automation coverage
